+++ crates/board/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 0.3.1-git
+## 0.4.0-git
+
+### Major
+
+- Change crypto API to mention `Keysize`, `BlockSize`, and `OutputSize`
+
+### Minor
+
+- Add `UnsupportedCrypto` for partially implemented crypto
 
 ### Patch